Moved checking the image homogeneity to images.py.
