+++ pylcogt/images.py
@@ -85,3 +85,10 @@ class Image(object):
 
 class InhomogeneousSetException(Exception):
     pass
+
+
+def check_image_homogeneity(images):
+    for attribute in ('nx', 'ny', 'ccdsum', 'epoch', 'site', 'instrument'):
+        if len({getattr(image, attribute) for image in images}) > 1:
+            raise InhomogeneousSetException('Images have different {}s'.format(attribute))
+    return images[0]
